feat(runbotics-orchestrator-ui): :sparkles: Protection against the creation of a variable with an existing name
