+++ runbotics/runbotics-orchestrator-ui/src/src-app/Actions/asana.actions.ts
@@ -36,6 +36,7 @@ const getAsanaActions: () => Record<string, IBpmnAction> = () => ({
                                 title: translate('Process.Details.Modeler.Actions.Common.VariableName'),
 
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
