+++ runbotics/runbotics-orchestrator-ui/src/src-app/Actions/beeOffice.actions.ts
@@ -74,6 +74,7 @@ const getBeeOfficeActions: () => Record<string, IBpmnAction> = () => ({
                                 title: translate('Process.Details.Modeler.Actions.Common.VariableName'),
 
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
@@ -142,6 +143,7 @@ const getBeeOfficeActions: () => Record<string, IBpmnAction> = () => ({
                                     'Process.Details.Modeler.Actions.Common.VariableName',
                                 ),
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
@@ -204,6 +206,7 @@ const getBeeOfficeActions: () => Record<string, IBpmnAction> = () => ({
                                     'Process.Details.Modeler.Actions.Common.VariableName',
                                 ),
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
@@ -262,6 +265,7 @@ const getBeeOfficeActions: () => Record<string, IBpmnAction> = () => ({
                             variableName: {
                                 title: translate('Process.Details.Modeler.Actions.Common.VariableName'),
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
@@ -330,6 +334,7 @@ const getBeeOfficeActions: () => Record<string, IBpmnAction> = () => ({
                                     'Process.Details.Modeler.Actions.Common.VariableName',
                                 ),
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
@@ -426,6 +431,7 @@ const getBeeOfficeActions: () => Record<string, IBpmnAction> = () => ({
                                     'Process.Details.Modeler.Actions.Common.VariableName',
                                 ),
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
@@ -494,6 +500,7 @@ const getBeeOfficeActions: () => Record<string, IBpmnAction> = () => ({
                                     'Process.Details.Modeler.Actions.Common.VariableName',
                                 ),
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
