+++ runbotics/runbotics-orchestrator-ui/src/src-app/Actions/file.actions.ts
@@ -185,6 +185,7 @@ const getFileActions = (): Record<string, IBpmnAction> => ({
                                 title: translate('Process.Details.Modeler.Actions.Common.VariableName'),
 
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
