+++ runbotics/runbotics-orchestrator-ui/src/src-app/Actions/general.actions.ts
@@ -123,6 +123,7 @@ const getGeneralActions: () => Record<string, IBpmnAction> = () => ({
                                 title: translate('Process.Details.Modeler.Actions.Common.VariableName'),
 
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: [],
