+++ runbotics/runbotics-orchestrator-ui/src/src-app/Actions/jira.actions.ts
@@ -49,6 +49,7 @@ const getJiraActions: () => Record<string, IBpmnAction> = () => ({
                                 title: translate('Process.Details.Modeler.Actions.Common.VariableName'),
 
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
