+++ runbotics/runbotics-orchestrator-ui/src/src-app/Actions/sharepointExcel.actions.ts
@@ -42,6 +42,7 @@ const getSharepointExcelActions: () => Record<string, IBpmnAction> = () => ({
                                 title: translate('Process.Details.Modeler.Actions.Common.VariableName'),
 
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
@@ -102,6 +103,7 @@ const getSharepointExcelActions: () => Record<string, IBpmnAction> = () => ({
                                 title: translate('Process.Details.Modeler.Actions.Common.VariableName'),
 
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
@@ -166,6 +168,7 @@ const getSharepointExcelActions: () => Record<string, IBpmnAction> = () => ({
                                 title: translate('Process.Details.Modeler.Actions.Common.VariableName'),
 
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
@@ -233,6 +236,7 @@ const getSharepointExcelActions: () => Record<string, IBpmnAction> = () => ({
                                 ),
 
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
@@ -322,6 +326,7 @@ const getSharepointExcelActions: () => Record<string, IBpmnAction> = () => ({
                                 ),
 
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
@@ -409,6 +414,7 @@ const getSharepointExcelActions: () => Record<string, IBpmnAction> = () => ({
                                 ),
 
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
@@ -462,6 +468,7 @@ const getSharepointExcelActions: () => Record<string, IBpmnAction> = () => ({
                                 ),
 
                                 type: 'string',
+                                pattern: '^[a-zA-Z0-9_]*$',
                             },
                         },
                         required: ['variableName'],
