+++ runbotics/runbotics-orchestrator-ui/src/src-app/hooks/useOptions.ts
@@ -1,0 +1,233 @@
+import { useMemo } from 'react';
+
+import useTranslations, {
+    translate as t,
+} from '#src-app/hooks/useTranslations';
+
+import { useSelector } from '#src-app/store';
+
+import useProcessVariables from './useProcessVariables';
+
+import { ActionVariableObject } from './useProcessVariables.types';
+
+const SERVICES = [
+    'environment.services.idt',
+    'environment.services.slice',
+    'environment.services.merge',
+    'environment.services.decrement',
+    'environment.services.increment',
+    'environment.services.get',
+    'environment.services.getLength',
+    'environment.services.getLastElemIdx',
+    'environment.services.jsonAsString',
+    'environment.services.objectAsBoolean',
+    'environment.services.stringObjectAsBoolean',
+    'environment.services.orObjectsAsBoolean',
+    'environment.services.andObjectsAsBoolean',
+    'environment.services.isEmptyString',
+    'environment.services.isEqualStr',
+    'environment.services.isEqual',
+    'environment.services.isNotEqual',
+    'environment.services.incrementAndCompare',
+    'environment.services.isBlank',
+    'environment.services.push',
+    'environment.services.concat',
+    'environment.services.splitAndPick',
+    'environment.services.parseJson',
+    'environment.services.csvToObject',
+    'environment.services.comaSeparatedToArray',
+    'environment.services.applyConverters',
+    'environment.services.withDefault',
+    'environment.services.getCurrentDate',
+    'environment.services.getPreviousWorkday',
+    'environment.services.checkCurrentTimeBefore',
+    'environment.services.checkCurrentTimeAfter',
+    'environment.services.checkCurrentTimeBetween',
+    'environment.services.getSecondsToNextDayTime',
+    'environment.services.getSecondsToSpecifiedHour',
+    'environment.services.getFirstWorkingDayOfCurrentMonth',
+    'environment.services.getFirstWorkingDayOfFutureMonth',
+    'environment.services.getLastWorkingDayOfCurrentMonth',
+    'environment.services.getLastWorkingDayOfFutureMonth',
+    'environment.services.checkFirstWorkingDayOfCurrentMonth',
+    'environment.services.checkLastWorkingDayOfCurrentMonth',
+    'environment.services.getSecondsToFirstWorkingDayOfMonth',
+    'environment.services.transformArraysToArray',
+    'environment.services.getCurrentShortMonthYear',
+    'environment.services.compareMaps',
+    'environment.services.diffMaps',
+    'environment.services.objFromArray',
+    'environment.services.readFromStorage',
+].map((service) => ({
+    value: `\${${service}}`,
+    label: `\${${service}}`,
+    name: service,
+    group: t(
+        'Process.Details.Modeler.Widgets.ElementAwareAutocomplete.Groups.Services'
+    ),
+}));
+
+const UTILS = [
+    'false',
+    'true',
+    'content.output',
+    'environment',
+    'environment.output',
+].map((util) => ({
+    value: `\${${util}}`,
+    label: `\${${util}}`,
+    name: util,
+    group: t(
+        'Process.Details.Modeler.Widgets.ElementAwareAutocomplete.Groups.Utils'
+    ),
+}));
+
+const VARIABLES = [
+    'tempFolder', 
+    'userEmail'
+].map((variable) => ({
+    label: variable,
+    value: variable,
+    group: t(
+        'Process.Details.Modeler.Widgets.ElementAwareAutocomplete.Groups.Variables'
+    ),
+}));
+
+const reduceList = (list: any[]) =>
+    list.reduce((previousValue, currentValue) => {
+        const newPrev = previousValue;
+        newPrev[currentValue.value] = currentValue;
+        return newPrev;
+    }, {});
+
+export interface Options {
+    [key: string]: {
+        label: string;
+        value: string;
+        group: string;
+        name: string;
+        actionId?: string;
+    };
+}
+
+const useOptions = () => {
+    const { selectedElement } = useSelector((state) => state.process.modeler);    
+    const { translate } = useTranslations();
+    const {
+        globalVariables,
+        inputActionVariables,
+        outputActionVariables,
+        attendedVariables,
+        loopVariables: scopedLoopVariables,
+    } = useProcessVariables(selectedElement?.parent?.id);
+    
+    const attendedProcessVariables = attendedVariables.map((variable) => ({
+        label: variable.name,
+        value: variable.name,
+        group: translate(
+            'Process.Details.Modeler.Widgets.ElementAwareAutocomplete.Groups.Variables'
+        ),
+    }));
+    
+    const extractOutputs = (outputVariables: ActionVariableObject[]) => {
+        const outputs = outputVariables.map((outputVariable) => ({
+            label: outputVariable.name,
+            value: outputVariable.name,
+            name: outputVariable.name,
+            ...(outputVariable?.actionId && { actionId: outputVariable?.actionId }),
+            group: translate(
+                'Process.Details.Modeler.Widgets.ElementAwareAutocomplete.Groups.Outputs'
+            ),
+        }));
+
+        const dollarOutputs = outputs.map((option) => ({
+            ...option,
+            label: `\${environment.output.${option.label}}`,
+            value: `\${environment.output.${option.value}}`,
+        }));
+        const hashOutputs = outputs.map((option) => ({
+            ...option,
+            label: `#{${option.label}}`,
+            value: `#{${option.value}}`,
+        }));
+
+        return [...dollarOutputs, ...hashOutputs];
+    };
+
+    const outputVariables = extractOutputs(outputActionVariables);
+
+    const groupedLocalVariables = inputActionVariables.map((variable) => ({
+        label: variable.name,
+        value: variable.name,
+        ...(variable?.actionId && { actionId: variable?.actionId }),
+        group: translate(
+            'Process.Details.Modeler.Widgets.ElementAwareAutocomplete.Groups.Variables'
+        ),
+    }));
+
+    const groupedLoopVariables = scopedLoopVariables.map((variable) => ({
+        label: variable.name,
+        value: variable.name,
+        group: translate(
+            'Process.Details.Modeler.Widgets.ElementAwareAutocomplete.Groups.Utils'
+        ),
+    }));
+
+    const groupedGlobalVariables = globalVariables.map((variable) => ({
+        label: variable.name,
+        value: variable.name,
+        group: translate(
+            'Process.Details.Modeler.Widgets.ElementAwareAutocomplete.Groups.Variables'
+        ),
+    }));
+
+    const options: Options = useMemo(() => {
+        let result = [];
+
+        const defaultOptions = [...SERVICES, ...UTILS];
+        result = [...defaultOptions, ...result];
+
+        const variables = [
+            ...groupedLocalVariables,
+            ...groupedGlobalVariables,
+            ...attendedProcessVariables,
+            ...VARIABLES,
+        ];
+        
+        const dollarVariables = variables.map((option) => ({
+            ...option,
+            label: `\${environment.variables.${option.value}}`,
+            value: `\${environment.variables.${option.value}}`,
+            name: option.value,
+        }));
+
+        const hashVariables = variables.map((option) => ({
+            ...option,
+            label: `#{${option.value}}`,
+            value: `#{${option.value}}`,
+            name: option.value,
+        }));
+
+        const loopVariables = groupedLoopVariables.map((option) => ({
+            ...option,
+            label: `\${${option.value}}`,
+            value: `\${${option.value}}`,
+            name: option.value,
+        }));
+
+        result = [
+            ...dollarVariables,
+            ...hashVariables,
+            ...outputVariables,
+            ...result,
+            ...loopVariables,
+        ];
+        
+        return reduceList(result);
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [selectedElement]);
+
+    return options;
+};
+
+export default useOptions;
