+++ runbotics/runbotics-orchestrator-ui/src/src-app/hooks/useProcessVariables.types.ts
@@ -3,6 +3,7 @@ export interface ActionVariableObject {
     name: string;
     value: string;
     scopeId?: string;
+    actionId?: string;
 }
 
 export interface LoopVariableObject extends ActionVariableObject {}
