+++ runbotics/runbotics-orchestrator-ui/src/src-app/translations/pl/action.json
@@ -17,6 +17,7 @@
     "Action.Details.Dialog.SaveAction": "Zapisz akcję: {{ script }}",
     "Action.Details.Dialog.LiveView": "PODGLĄD NA ŻYWO",
 
+    "Process.BuildView.Modeler.Widgets.VariableNameIsTaken": "Nazwa zmiennej już istnieje",
     "Process.BuildView.Modeler.Widgets.FieldTemplate.AssignTheOutputToVariable": "Przypisz dane wyjściowe do zmiennej",
     "Process.BuildView.Modeler.Widgets.FieldTemplate.IsARequiredProperty": "Ta właściwość jest wymagana",
     "Process.BuildView.Modeler.Widgets.FieldTemplate.ShouldMatchPattern": "Użyto niedozwolonych znaków",
