+++ runbotics/runbotics-orchestrator-ui/src/src-app/views/process/ProcessBuildView/Modeler/ActionFormPanel/renderers/JSONSchemaFormRenderer.tsx
@@ -14,7 +14,7 @@ import _ from 'lodash';
 
 import useDebounce from '#src-app/hooks/useDebounce';
 import { translate } from '#src-app/hooks/useTranslations';
-import { useDispatch } from '#src-app/store';
+import { useDispatch, useSelector } from '#src-app/store';
 
 import { ModelerError, processActions } from '#src-app/store/slices/Process';
 
@@ -68,6 +68,7 @@ const JSONSchemaFormRenderer: FC<FormPropsExtended> = (props) => {
     const isFormDirtyRef = useRef(isFormDirty);
     const editModeRef = useRef(editMode);
     const debouncedForm = useDebounce<FormState>(formState, DEBOUNCE_TIME);
+    const { selectedElement } = useSelector(state => state.process.modeler);
 
     const handleSubmit = (e: any, nativeEvent?: FormEvent<HTMLFormElement>) => {
         setEditMode(false);
@@ -131,6 +132,7 @@ const JSONSchemaFormRenderer: FC<FormPropsExtended> = (props) => {
                             formData={formState.formData}
                             onChange={handleChange}
                             FieldTemplate={FieldTemplate}
+                            formContext={{ selectedElement }}
                         >
                             <Button type="submit" style={{ display: 'none' }} />
                         </Form>
